refactor: decompose azure_devops_pipeline.go — extract log processing subsystem
Move 16 pure stream-processing symbols to azure_devops_pipeline_logs.go.
Same package, same receiver type (*adoManager), zero exported API changes.

Symbols moved: logFilterOptions, filterResult, processLogContent,
streamRegexFilter, logFilterState (type + constructor + 5 methods),
streamPagination, streamHead, streamTail, scanLog, assembleTail.

Kept in pipeline.go: adoGetPipelineLogs, adoGetTaskLog,
listPipelineLogs, fetchPipelineLogContent (ADO orchestration).

Closes #100
